+++ src/main/interfaces/ILoggerOption.ts
@@ -11,6 +11,43 @@ export enum ELoggerLevel {
   DEBUG = 4
 }
 
+export enum ELoggerLevelNames {
+  'INFO',
+  'WARN',
+  'ERROR',
+  'FATAL',
+  'DEBUG'
+}
+
+// color for text or colored background
+export enum ELoggerLevelBaseColors {
+  '#cc80ff',
+  '#ff8a1c',
+  '#ff4a4a',
+  '#ffffff',
+  '#555555',
+}
+
+// color for text on colored background
+export const ELoggerLevelAlternateColors = [
+  '#000000',
+  '#000000',
+  '#000000',
+  '#ff0000',
+  '#D4D4D4'
+]
+
+export interface ILoggerFileProperties {
+  enable: boolean;
+  logFolderPath: string;
+  enableLatestLog?: boolean;
+  enableDebugLog?: boolean;
+  enableErrorLog?: boolean;
+  enableFatalLog?: boolean;
+  generateHTMLLog?: boolean;
+  compressLogFilesAfterNewExecution?: boolean;
+}
+
 export interface ILoggerOptions {
   defaultLevel?: ELoggerLevel;
   prefix?: string;
@@ -18,4 +55,5 @@ export interface ILoggerOptions {
   coloredBackground?: boolean;
   allLineColored?: boolean;
   disableFatalCrash?: boolean;
+  fileProperties?: ILoggerFileProperties
 }